+++ Realistic Hololens Rendering/Content/Shaders/Camera Testing/VprtVertexShader.hlsl
@@ -10,15 +10,15 @@ cbuffer ViewProjectionConstantBuffer : register(b1)
 
 struct VertexShaderInput
 {
-	min16float3 position : POSITION;
-	min16float2 uv : TEXCOORD0;
+	float3 position : POSITION;
+	float2 uv : TEXCOORD0;
 	uint instanceId : SV_InstanceID;
 };
 
 struct VertexShaderOutput
 {
-	min16float4 position : SV_POSITION;
-	min16float2 uv : TEXCOORD0;
+	float4 position : SV_POSITION;
+	float2 uv : TEXCOORD0;
 	uint renderTargetId : SV_RenderTargetArrayIndex;
 };
 
@@ -27,7 +27,7 @@ VertexShaderOutput main(VertexShaderInput input)
 	VertexShaderOutput output;
 	float4 position = float4(input.position, 1.0f);
 	position = mul(mul(position, model), viewProjection[input.instanceId % 2]);
-	output.position = (min16float4)position;
+	output.position = (float4)position;
 	output.uv = input.uv;
 	output.renderTargetId = input.instanceId % 2;
 	return output;
