+++ Realistic Hololens Rendering/Content/Shaders/Cube Rendering/VprtVertexShader.hlsl
@@ -16,16 +16,16 @@ cbuffer CameraConstantBuffer : register(b2)
 
 struct VertexShaderInput
 {
-	min16float3 Position : POSITION;
-	min16float3 Normal : NORMAL;
+	float3 Position : POSITION;
+	float3 Normal : NORMAL;
 	uint InstanceId : SV_InstanceID;
 };
 
 struct VertexShaderOutput
 {
-	min16float4 Position : SV_POSITION;
-	min16float4 Normal : NORMAL;
-	min16float2 UV : TEXCOORD;
+	float4 Position : SV_POSITION;
+	float4 Normal : NORMAL;
+	float2 UV : TEXCOORD;
 	uint RenderTargetId : SV_RenderTargetArrayIndex;
 };
 
@@ -35,16 +35,16 @@ VertexShaderOutput main(VertexShaderInput input)
 	float4 position = float4(input.Position, 1.0f);
 	float4 worldPosition = mul(position, VertexModel);
 	position = mul(worldPosition, ViewProjection[input.InstanceId % 6]);
-	output.Position = (min16float4)position;
+	output.Position = (float4)position;
 	float4 normal = float4(input.Normal, 0.0f);
 	normal = mul(normal, NormalModel);
-	output.Normal = (min16float4)normal;
+	output.Normal = (float4)normal;
 	output.RenderTargetId = input.InstanceId % 6;
 	float4 cameraPosition = mul(worldPosition, CameraViewProjection);
 	float w = cameraPosition.w;
 	cameraPosition /= w;
 	float2 cameraUV = cameraPosition.xy;
 	cameraUV = (cameraUV + 1.0f) / 2.0f;
-	output.UV = w > 0 ? (min16float2)cameraUV : min16float2(-1.0, -1.0);
+	output.UV = w > 0 ? (float2)cameraUV : float2(-1.0, -1.0);
 	return output;
 }
