+++ Realistic Hololens Rendering/Content/Shaders/Mesh Projection/DepthPrepassPixelShader.hlsl
@@ -1,6 +1,6 @@
 struct PixelShaderInput
 {
-	min16float4 Position : SV_Position;
+	float4 Position : SV_Position;
 };
 
 void main(PixelShaderInput input)
