+++ Realistic Hololens Rendering/Content/Shaders/Mesh Projection/DepthPrepassVertexShader.hlsl
@@ -16,19 +16,19 @@ cbuffer CameraConstantBuffer : register(b2)
 
 struct VertexShaderInput
 {
-	min16float3 Position : Position;
-	min16float3 Normal : Normal;
+	float3 Position : Position;
+	float3 Normal : Normal;
 };
 
 struct VertexShaderOutput
 {
-	min16float4 Position : SV_Position;
+	float4 Position : SV_Position;
 };
 
 VertexShaderOutput main(VertexShaderInput input)
 {
 	VertexShaderOutput output;
-	min16float4 position = min16float4(input.Position, 1.0);
+	float4 position = float4(input.Position, 1.0);
 	position = mul(mul(position, VertexModel), CameraViewProjection);
 	output.Position = position;
 	return output;
