+++ Realistic Hololens Rendering/Content/Shaders/Mesh Projection/PixelShader.hlsl
@@ -2,14 +2,14 @@
 
 struct PixelShaderInput
 {
-	min16float4 Position : SV_POSITION;
-	min16float3 WorldSpace : POSITION;
+	float4 Position : SV_POSITION;
+	float3 WorldSpace : POSITION;
 };
 
 struct PixelShaderOutput
 {
-	min16float4 Color : SV_Target0;
-	// min16float2 QualityAndTime : SV_Target1;
+	float4 Color : SV_Target0;
+	// float2 QualityAndTime : SV_Target1;
 };
 
 cbuffer CameraConstantBuffer : register(b2)
@@ -30,11 +30,11 @@ SamplerState TextureSamplerState
 	Filter = MIN_MAG_MIP_POINT;
 };
 
-min16float3 YuvToRgb(min16float2 textureUV);
+float3 YuvToRgb(float2 textureUV);
 
 PixelShaderOutput main(PixelShaderInput input)
 {
-	min16float4 lightSpace = mul(min16float4(input.WorldSpace, 1.0), CameraViewProjection);
+	float4 lightSpace = mul(float4(input.WorldSpace, 1.0), CameraViewProjection);
 	if (lightSpace.w < 0.0)
 	{
 		discard;
@@ -56,12 +56,12 @@ PixelShaderOutput main(PixelShaderInput input)
 	}
 
 	PixelShaderOutput output;
-	output.Color = min16float4(YuvToRgb(lightSpace.xy), 1.0);
-	// output.QualityAndTime = min16float2(0.0, 0.0);
+	output.Color = float4(YuvToRgb(lightSpace.xy), 1.0);
+	// output.QualityAndTime = float2(0.0, 0.0);
 	return output;
 }
 
-min16float3 YuvToRgb(min16float2 textureUV)
+float3 YuvToRgb(float2 textureUV)
 {
 	int3 location = int3(0, 0, 0);
 	location.x = (int)(1408 * (textureUV.x));
@@ -74,7 +74,7 @@ min16float3 YuvToRgb(min16float2 textureUV)
 	int r = (298 * c + 409 * e + 128) >> 8;
 	int g = (298 * c - 100 * d - 208 * e + 128) >> 8;
 	int b = (298 * c + 516 * d + 128) >> 8;
-	min16float3 rgb = float3(0.0f, 0.0f, 0.0f);
+	float3 rgb = float3(0.0f, 0.0f, 0.0f);
 	rgb.x = max(0, min(255, r));
 	rgb.y = max(0, min(255, g));
 	rgb.z = max(0, min(255, b));
