+++ Realistic Hololens Rendering/Content/Shaders/Mesh Projection/VertexShader.hlsl
@@ -11,21 +11,21 @@ cbuffer ViewProjectionConstantBuffer : register(b1)
 
 struct VertexShaderInput
 {
-	min16float3 Position : Position;
-	min16float3 Normal : Normal;
+	float3 Position : Position;
+	float3 Normal : Normal;
 };
 
 struct VertexShaderOutput
 {
-	min16float4 Position : SV_Position;
-	min16float3 WorldSpace : Position;
+	float4 Position : SV_Position;
+	float3 WorldSpace : Position;
 };
 
 VertexShaderOutput main(VertexShaderInput input)
 {
 	VertexShaderOutput output;
-	output.Position = min16float4(0.0, 0.0, 0.0, 0.0);
-	min16float4 position = min16float4(input.Position, 1.0);
+	output.Position = float4(0.0, 0.0, 0.0, 0.0);
+	float4 position = float4(input.Position, 1.0);
 	position = mul(position, VertexModel);
 	output.WorldSpace = position.xyz;
 	return output;
