+++ Realistic Hololens Rendering/Content/Shaders/Mesh Rendering/PixelShader.hlsl
@@ -1,7 +1,7 @@
 struct PixelShaderInput
 {
-	min16float4 Position : SV_Position;
-	min16float2 UV : TexCoord;
+	float4 Position : SV_Position;
+	float2 UV : TexCoord;
 };
 
 Texture2D<float4> Color : register(t0);
@@ -12,7 +12,7 @@ SamplerState TextureSamplerState
 	Filter = MIN_MAG_MIP_POINT;
 };
 
-min16float4 main(PixelShaderInput input) : SV_Target
+float4 main(PixelShaderInput input) : SV_Target
 {
 	return Color.Sample(TextureSamplerState, input.UV);
 }
