+++ Realistic Hololens Rendering/Content/Shaders/Mesh Updating/VertexShader.hlsl
@@ -11,20 +11,20 @@ cbuffer ViewProjectionConstantBuffer : register(b1)
 
 struct VertexShaderInput
 {
-	min16float3 Position : Position;
-	min16float3 Normal : Normal;
+	float3 Position : Position;
+	float3 Normal : Normal;
 };
 
 struct VertexShaderOutput
 {
-	min16float4 Position : SV_Position;
-	min16float2 UV : TexCoord;
+	float4 Position : SV_Position;
+	float2 UV : TexCoord;
 };
 
 VertexShaderOutput main(VertexShaderInput input)
 {
 	VertexShaderOutput output;
-	output.Position = min16float4(0.0, 0.0, 0.0, 0.0);
-	output.UV = min16float2(0.0, 0.0);
+	output.Position = float4(0.0, 0.0, 0.0, 0.0);
+	output.UV = float2(0.0, 0.0);
 	return output;
 }
