+++ Realistic Hololens Rendering/Content/Shaders/Texture Debug/PixelShader.hlsl
@@ -1,7 +1,7 @@
 struct PixelShaderInput
 {
-	min16float4 Position : SV_Position;
-	min16float2 UV : TexCoord;
+	float4 Position : SV_Position;
+	float2 UV : TexCoord;
 };
 
 Texture2D<float4> Color : register(t0);
@@ -11,7 +11,7 @@ SamplerState TextureSamplerState
 	Filter = MIN_MAG_MIP_LINEAR;
 };
 
-min16float4 main(PixelShaderInput input) : SV_Target
+float4 main(PixelShaderInput input) : SV_Target
 {
-	return min16float4(Color.Sample(TextureSamplerState, input.UV).rgb / 2.0, 1.0);
+	return float4(Color.Sample(TextureSamplerState, input.UV).rgb / 2.0, 1.0);
 }
