+++ Realistic Hololens Rendering/Content/Shaders/Texture Debug/VertexShader.hlsl
@@ -10,22 +10,22 @@ cbuffer ViewProjectionConstantBuffer : register(b1)
 
 struct VertexShaderInput
 {
-	min16float3 Position : Position;
-	min16float2 UV : TexCoord;
+	float3 Position : Position;
+	float2 UV : TexCoord;
 	uint InstanceID : SV_InstanceID;
 };
 
 struct VertexShaderOutput
 {
-	min16float4 Position : SV_Position;
-	min16float2 UV : TexCoord;
+	float4 Position : SV_Position;
+	float2 UV : TexCoord;
 	uint InstanceID : SV_RenderTargetArrayIndex;
 };
 
 VertexShaderOutput main(VertexShaderInput input)
 {
 	VertexShaderOutput output;
-	output.Position = min16float4(input.Position.xy * Bounds.zw + Bounds.xy, 0.0, 1.0);
+	output.Position = float4(input.Position.xy * Bounds.zw + Bounds.xy, 0.0, 1.0);
 	output.UV = input.UV;
 	output.InstanceID = input.InstanceID;
 	return output;
